requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
docopt==0.6.2
geographiclib==2.0
geopy==2.4.1
idna==3.10
packaging==24.2
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
requests==2.32.3
setuptools==76.0.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0

